requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-ai-documentintelligence==1.0.2
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.33.0
azure-cosmos==4.9.0
azure-functions==1.23.0
azure-functions-durable==1.2.10
azure-identity==1.21.0
azure-storage-blob==12.25.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorclass==2.2.2
compressed-rtf==1.0.7
cryptography==44.0.0
distro==1.9.0
easygui==0.98.3
ebcdic==1.1.1
et_xmlfile==2.0.0
extract-msg==0.54.1
frozenlist==1.6.0
furl==2.1.4
google-auth==2.40.3
google-genai==1.39.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
jiter==0.9.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.1.9
lxml==5.4.0
MarkupSafe==3.0.2
mcp==1.15.0
msal==1.32.2
msal-extensions==1.3.1
msoffcrypto-tool==5.4.2
msrest==0.7.1
multidict==6.4.3
numpy==2.2.5
oauthlib==3.2.2
olefile==0.47
oletools==0.60.2
openai==1.76.0
openpyxl==3.1.5
orderedmultidict==1.0.1
pandas==2.2.3
pcodedmp==1.2.6
pillow==11.2.1
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
red-black-tree-mod==1.22
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
RTFDE==0.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.3
starlette==0.46.2
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
Werkzeug==3.1.3
XlsxWriter==3.2.3
yarl==1.20.0
git+https://github.com/estyle-llm/i-style.git@1.22.4